feat(fields): adds a select field and updates some styles
