+++ auto-decider/src/App.css
@@ -41,7 +41,7 @@ input {
   border: 1px solid white;
   padding: 10px 15px;
   margin-bottom: 10px;
-  font-size: 14px;
+  font-size: 16px;
 }
 
 label {
@@ -51,7 +51,7 @@ label {
   margin-bottom: 13px;
   margin-top: 20px;
   color: white;
-  font-size: 14px;
+  font-size: 16px;
   font-weight: 200;
 }
