+++ auto-decider/src/Options/CarOptions.js
@@ -1,0 +1,9 @@
+const RowOptions = [
+    { title: 'Yes', value: true, id: 1 },
+    { title: 'No', value: false, id: 2 },
+    { title: 'Not sure', value: null, id: 3 },
+]
+
+export {
+    RowOptions
+}
